+++ src/common/shared/calendar/type.ts
@@ -1,3 +1,9 @@
+export type TCalendarValue = number | Date;
+
+export type TDateType = 'selected' | 'disabled' | 'start' | 'centre' | 'end' | '';
+
+export type TCalendarType = 'single' | 'multiple' | 'range';
+
 export interface TDate {
   date: Date;
   day: number;
@@ -6,9 +12,3 @@ export interface TDate {
   prefix?: string;
   suffix?: string;
 }
-
-export type TCalendarValue = number | Date;
-
-export type TDateType = 'selected' | 'disabled' | 'start' | 'centre' | 'end' | '';
-
-export type TCalendarType = 'single' | 'multiple' | 'range';