+++ src/common/src/superComponent.ts
@@ -1,11 +1,5 @@
 export interface ComponentsOptionsType extends WechatMiniprogram.Component.ComponentOptions {
-  styleIsolation?:
-  | 'isolated'
-  | 'apply-shared'
-  | 'shared'
-  | 'page-isolated'
-  | 'page-apply-shared'
-  | 'page-shared';
+  styleIsolation?: 'isolated' | 'apply-shared' | 'shared' | 'page-isolated' | 'page-apply-shared' | 'page-shared';
 }
 
 export interface RelationsOptions {
@@ -14,8 +8,8 @@ export interface RelationsOptions {
 
 export interface SuperComponent<D = {}, P = {}, M = {}>
   extends WechatMiniprogram.Component.Lifetimes,
-  WechatMiniprogram.Component.OtherOption,
-  WechatMiniprogram.Component.InstanceMethods<D> {
+    WechatMiniprogram.Component.OtherOption,
+    WechatMiniprogram.Component.InstanceMethods<D> {
   properties: P;
   // 模拟 properties 被代理到 data 上的声明
   data: D;
@@ -25,6 +19,7 @@ export interface SuperComponent<D = {}, P = {}, M = {}>
   [x: string]: any;
 }
 
+// eslint-disable-next-line no-redeclare, @typescript-eslint/no-unused-vars
 export class SuperComponent<D = {}, P = {}, M = {}> {
   readonly app: any;