+++ src/count-down/count-down.ts
@@ -28,6 +28,8 @@ export default class CountDown extends SuperComponent {
 
   timeoutId: null | number = null;
 
+  isInitialTime = false;
+
   lifetimes = {
     detached() {
       if (this.timeoutId) {
@@ -58,9 +60,11 @@ export default class CountDown extends SuperComponent {
       this.remain = this.properties.time;
       this.updateTime(this.remain);
 
-      if (this.properties.autoStart) {
+      if (this.properties.autoStart && this.remain > 0) {
         this.start();
       }
+
+      this.isInitialTime = true;
     },
 
     getTime(): number {
@@ -80,9 +84,10 @@ export default class CountDown extends SuperComponent {
         formattedTime: timeText.replace(/:/g, ' : '),
       });
 
-      if (remain === 0) {
+      if (remain === 0 && (this.counting || this.isInitialTime)) {
         this.pause();
         this.triggerEvent('finish');
+        this.counting = false;
       }
     },