+++ src/dialog/dialog.ts
@@ -1,7 +1,8 @@
 import { SuperComponent, wxComponent } from '../common/src/index';
 import config from '../common/config';
 import props from './props';
-import { isObject, toCamel } from '../common/utils';
+import { toCamel } from '../common/utils';
+import { isObject } from '../common/validator';
 import useCustomNavbar from '../mixins/using-custom-navbar';
 
 const { prefix } = config;
@@ -109,24 +110,30 @@ export default class Dialog extends SuperComponent {
 
     onConfirm() {
       this.triggerEvent('confirm');
+
       if (this._onConfirm) {
-        this._onConfirm();
+        this._onConfirm({ trigger: 'confirm' });
         this.close();
       }
     },
 
     onCancel() {
-      this.triggerEvent('close', { trigger: 'cancel' });
+      const trigger = { trigger: 'cancel' };
+
       this.triggerEvent('cancel');
+      this.triggerEvent('close', trigger);
 
       if (this._onCancel) {
-        this._onCancel();
+        this._onCancel(trigger);
         this.close();
       }
     },
 
     onClose() {
-      this.triggerEvent('close', { trigger: 'close-btn' });
+      const trigger = { trigger: 'close-btn' };
+
+      this.triggerEvent('close', trigger);
+      this._onCancel?.(trigger);
       this.close();
     },
 
@@ -135,11 +142,15 @@ export default class Dialog extends SuperComponent {
     },
 
     overlayClick() {
+      this.triggerEvent('overlay-click');
+
       if (this.properties.closeOnOverlayClick) {
-        this.triggerEvent('close', { trigger: 'overlay' });
+        const trigger = { trigger: 'overlay' };
+
+        this.triggerEvent('close', trigger);
+        this._onCancel?.(trigger);
         this.close();
       }
-      this.triggerEvent('overlay-click');
     },
 
     onActionTap(index: number) {