+++ src/upload/type.ts
@@ -8,6 +8,14 @@ import { ImageProps } from '../image/index';
 
 export interface TdUploadProps {
   /**
+   * 添加按钮
+   * @default true
+   */
+  addBtn?: {
+    type: BooleanConstructor;
+    value?: boolean;
+  };
+  /**
    * 添加按钮内容
    */
   addContent?: {
@@ -94,6 +102,14 @@ export interface TdUploadProps {
   mediaType?: {
     type: ArrayConstructor;
     value?: Array<MediaType>;
+  };
+  /**
+   * 移除按钮
+   * @default true
+   */
+  removeBtn?: {
+    type: BooleanConstructor;
+    value?: boolean;
   };
   /**
    * 自定义上传方法